pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pusht619"
version = "0.1.0"
description = "CSCI 619 class project"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "CSCI 619" }]
dependencies = [
    "numpy",
    "jaxsim",
    "matplotlib",
    "python-dateutil",
    "rod",
    "mujoco",
    "lxml",
    "mediapy",
    "imageio[ffmpeg]",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.urls]
Homepage = "https://example.com/csci619_class_project"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
addopts = "-ra -q"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
